requirements.txt
gradio>=4.40.0
requests>=2.31.0
aiohttp>=3.8.0
python-multipart>=0.0.6
websockets>=11.0
pillow>=9.0.0
pydantic>=2.0.0
uvicorn>=0.20.0
fastapi>=0.100.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0